requirements.txt
numpy==1.18.1
numpydoc==0.9.2
pandas==1.0.3
scikit-image==0.16.2
scikit-learn==0.22.1
sklearn-pandas==1.8.0